+++ clients/dashboard/dashboard/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
+    "@material-ui/core": "^4.12.3",
     "@mui/icons-material": "^5.2.4",
     "@mui/material": "^5.2.4",
     "@testing-library/jest-dom": "^5.16.1",
