+++ clients/dashboard/dashboard/src/components/waitersDialog.js
@@ -4,7 +4,7 @@ import ListItem from "@mui/material/ListItem";
 import ListItemText from "@mui/material/ListItemText";
 import DialogTitle from "@mui/material/DialogTitle";
 import Dialog from "@mui/material/Dialog";
-import { getWaiters } from "../api";
+import { getWaiters } from "../network/api";
 
 const dialogTitle = "Choose Waiter";
