--- clients/dashboard/dashboard/src/config.json
@@ -1,4 +1,0 @@
-{
-    "port": 3000, 
-    "host": "https://service-everywhere.herokuapp.com/"
-}
