+++ clients/dashboard/dashboard/src/network/config.json
@@ -1,0 +1,4 @@
+{
+  "port": 3000,
+  "host": "https://service-everywhere.herokuapp.com/"
+}
